backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
joblib
psycopg2-binary